enhancement(http): impl `Display` directly on `Route` & `RequestReactionType`
Having `Display` be behind a seperate method doesn't hold much value.
