+++ http/src/request/base.rs
@@ -194,7 +194,7 @@ impl Request {
             form: None,
             headers: None,
             method: route.method(),
-            path: route.display().to_string(),
+            path: route.to_string(),
             ratelimit_path: route.path(),
             use_authorization_token: true,
         }
