+++ http/src/routing/mod.rs
@@ -1,5 +1,4 @@
 mod route;
-mod route_display;
 
-pub use self::{route::Route, route_display::RouteDisplay};
+pub use self::route::Route;
 pub use twilight_http_ratelimiting::request::{Path, PathParseError, PathParseErrorType};
